refactor(cli): retire legacy __main__ entry points
python -m maildb now dispatches to the Typer app and defaults to
the serve subcommand. The standalone ingest entrypoint is replaced
by maildb ingest run/status/reset/migrate. Bundles lint/typing cleanup
from the Typer introduction.
